003 at:  003 caservice 생성

+++ src/main/java/com/mysite/cuffee/cart/repository/CartRepository.java
@@ -3,5 +3,6 @@ package com.mysite.cuffee.cart.repository;
 import com.mysite.cuffee.cart.entity.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartRepository extends JpaRepository<Cart,Long> {
+
+public interface CartRepository extends JpaRepository<Cart,Integer> {
 }
